Replaced 0 valules by NAs when it's the case, so that we don't create artificial zeros that lower the indicator. Also, did the QCC plot on our data.
Results : We do have an alert during the symptomatic phase, good thing. But afterwards the indicator is very normal. However, we do have an alert just before the death, so even if it's not soon enough to do someting to prevent it, it's good to see that we can see the problem when it occurs.